requirements.txt
social-auth-app-django
django-widget-tweaks==1.4.1
psycopg2==2.7.3.1